load data as linear RGB encoding
Technically, this is probably wrong, as the images are likely in sRGB
colorspace. However, it gives a better match to the results (visually
inspected) of operations with "Color Inspector 3D" by Kai Uwe Barthel.
Furthermore, apparently [it is not specified what colorspace browsers
use to draw images in the canvas
element](https://wiki.whatwg.org/wiki/CanvasColorSpace).